requirements.txt
protobuf==6.31.1
pydantic==2.11.7
pydantic_settings==2.10.1
google-cloud-aiplatform==1.103.0
google-cloud-storage==3.2.0
google-cloud-bigquery==3.35.0